Integrate with Sqitch
I like postgres. I will be using postgres in this project. I want to
make sure that I take as much advantage of the DB as I want to, given
the purpose of this, so I am not integrating with Ecto.

I wanted the freedom to play with Postgres as much as possible, so
sqitch allows that.